fix bug in register

+++ src/controllers/auth.js
@@ -124,10 +124,7 @@ module.exports = {
             const isExist = await User.findOne({ where: { phone_number } })
             // console.log(isExist === null)
             if (isExist) {
-              if (req.file) {
-                fs.unlinkSync('assets/uploads/' + req.file.filename)
-                return response(res, 'Error phone number has been registered, please login with it,', {}, 400, false)
-              }
+              fs.unlinkSync('assets/uploads/' + req.file.filename)
               return response(res, 'Error phone number has been registered, please login with it,', {}, 400, false)
             } else {
               console.log(data)
@@ -141,7 +138,7 @@ module.exports = {
               phone_number
             }
             const isExist = await User.findOne({ where: { password } })
-            // console.log(isExist === null)
+            console.log(isExist)
             if (isExist) {
               return response(res, 'Error phone number has been registered, please login with it,', {}, 400, false)
             } else {
@@ -151,11 +148,13 @@ module.exports = {
             }
           }
         } else {
-          return response(res, 'please input phone number', 400, false);
+          return response(res, 'please input phone number', 400, false)
         }
       } catch (e) {
-        fs.unlinkSync('assets/uploads/' + req.file.filename)
-        return response(res, e.message, {}, 500, false)
+        if (req.file) {
+          fs.unlinkSync('assets/uploads/' + req.file.filename)
+          return response(res, e.message, {}, 500, false)
+        }
       }
     })
   },
